requirements.txt
pytest >= 8.3.2
requests~=2.31.0
jsonschema~=4.23.0
allure-pytest>=2.13.5
selene>=2.0.0rc9
webdriver-manager>=4.0.2

